+++ src/apis/leaderboard.request.ts
@@ -1,5 +1,5 @@
-import { LeaderboardListDto } from '@/types/leaderboard.type';
 import { PATHS } from '@/constants';
+import { LeaderboardListDto } from '@/types/leaderboard.type';
 import { instance } from './instance.request';
 
 export const leaderboardList = async ({
